+++ MarkedMap.cpp
@@ -6,9 +6,9 @@ MarkedMap::MarkedMap(const std::string &filename) : Map(filename)
 
 std::pair<int, int> MarkedMap::getHeroPosition() const
 {
-	for (int i = 0; i < map.size(); i++)
+	for (size_t i = 0; i < map.size(); i++)
 	{
-		for (int j = 0; j < map[i].size(); j++)
+		for (size_t j = 0; j < map[i].size(); j++)
 		{
 			if (map[i][j] == 'H')
 			{
@@ -24,9 +24,9 @@ std::pair<int, int> MarkedMap::getHeroPosition() const
 std::vector<std::pair<int, int>> MarkedMap::getMonsterPositions(char c) const
 {
 	std::vector<std::pair<int, int>> positions;
-	for (int i = 0; i < map.size(); i++)
+	for (size_t i = 0; i < map.size(); i++)
 	{
-		for (int j = 0; j < map[i].size(); j++)
+		for (size_t j = 0; j < map[i].size(); j++)
 		{
 			if (map[i][j] == c)
 			{
